requirements.txt
flask==1.1.1
requests==2.22.0
pyyaml==3.13